pyproject.toml
[tool.poetry]
name = "hw-24-drf"
version = "0.1.0"
description = ""
authors = ["Olga Lebedeva <olga.s.lebedeva@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
ipython = "^8.22.2"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
django-filter = "^24.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.4"
drf-yasg = "^1.21.7"
setuptools = "^69.2.0"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
stripe = "^8.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
